eliminate one of the data races on access to node data
This particular data race hits us a lot in CI testing. The
underlying issue is that access to the contents of m.nodeMap
is controlled by a mutex, but rawSendMsgPacket is using the
mutex only to grab the pointer -- then accessing the contents
of the pointer without the mutex held. This means a data race
if something else grabs the mutex and modifies the contents
of the pointer as well.

In this particular case, access is read-only, so we copy the
pointed-to thing into a local. Less efficient, but also less
of a data race.